ENH: Implement abstract classes
