+++ mitoviz/__init__.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 # -*- coding: UTF-8 -*-
 # Created by Roberto Preste
-from mitoviz.mitoviz import plot_df, plot_table, plot_vcf
+from mitoviz.mitoviz import plot_df, plot_table, plot_vcf  # noqa
 
 __author__ = """Roberto Preste"""
 __email__ = "robertopreste@gmail.com"
