+++ mitoviz/classes/__init__.py
@@ -1,0 +1,3 @@
+#!/usr/bin/env python
+# -*- coding: UTF-8 -*-
+# Created by Roberto Preste
