+++ mitoviz/tests/constants.py
@@ -76,6 +76,7 @@ SAMPLE_CUSTOM_DF = pd.DataFrame(
     {"position": [750, 1438, 2706, 3270, 4216, 7028, 8935, 9389, 11251, 12633],
      "reference": ["A", "A", "A", "CT", "T", "C", "C", "A", "A", "C"],
      "alternate": ["G", "G", "G", "C", "C", "T", "T", "G", "G", "A"],
-     "het_frac": [0.998, 0.232, 0.015, 0.66, 1.0, 0.420, 0.899, 1.0, 0.999, 0.998],
+     "het_frac": [0.998, 0.232, 0.015, 0.66, 1.0, 0.420, 0.899, 1.0, 0.999,
+                  0.998],
      "samplename": ["HG00420"] * 10}
 )
